Fixes circular dependency
