+++ test/test_specific_conv.py
@@ -1,6 +1,6 @@
 import unittest
 from tinygrad.tensor import Tensor
-from tinygrad.helpers import dtypes
+from tinygrad.dtypes import dtypes
 from tinygrad.lazy import Device
 # similar to test/external/external_test_gpu_ast.py, but universal
