+++ tinygrad/DType.py
@@ -1,0 +1,10 @@
+from typing import NamedTuple, Optional
+class DType(NamedTuple):
+  priority: int  # this determines when things get upcasted
+  itemsize: int
+  name: str
+  np: Optional[type]  # TODO: someday this will be removed with the "remove numpy" project
+  sz: int = 1
+  def __repr__(self): return f"dtypes.{self.name}"
+  @property
+  def key(self) -> str: return (self.name)
