+++ tinygrad/helpers.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
-import os, functools, platform, time, re
-from weakref import KeyedRef, ref
+import platform, time, re
 from _weakref import _remove_dead_weakref # type: ignore
 import inspect
-import numpy as np
-from typing import Dict, Tuple, Union, List, NamedTuple, Final, Iterator, ClassVar, Optional, Callable, Any
+from typing import Tuple, Union, List, Iterator, ClassVar, Any
+from tinygrad.GRAPH_MACROS import getenv
 from math import prod # noqa: F401 # pylint:disable=unused-import
 
 ShapeType = Tuple[int, ...]
@@ -27,9 +26,6 @@ def flatten(l:Iterator): return [item for sublist in l for item in sublist]
 def mnum(i) -> str: return str(i) if i >= 0 else f"m{-i}"
 def fromimport(mod, frm): return getattr(__import__(mod, fromlist=[frm]), frm)
 
-@functools.lru_cache(maxsize=None)
-def getenv(key, default=0): return type(default)(os.getenv(key, default))
-
 class Context:
   def __init__(self, **kwargs): self.pvars = kwargs
   def __enter__(self): ContextVar.ctx_stack.append({ **self.pvars, **{ key: ContextVar.ctx_stack[-1][key] for key in ContextVar.ctx_stack[-1].keys() if key not in self.pvars } })
@@ -50,7 +46,6 @@ class ContextVar:
   def value(self) -> int: return ContextVar.ctx_stack[-1][self.key] if self.key in ContextVar.ctx_stack[-1] else self.initial_value
 
 DEBUG, IMAGE = ContextVar("DEBUG", 0), ContextVar("IMAGE", 0)
-GRAPH, PRUNEGRAPH, GRAPHPATH = getenv("GRAPH", 0), getenv("PRUNEGRAPH", 0), getenv("GRAPHPATH", "/tmp/net")
 
 class Timing(object):
   def __init__(self, prefix="", on_exit=None, enabled=True): self.prefix, self.on_exit, self.enabled = prefix, on_exit, enabled
@@ -58,103 +53,3 @@ class Timing(object):
   def __exit__(self, exc_type, exc_val, exc_tb):
     self.et = time.perf_counter_ns() - self.st
     if self.enabled: print(f"{self.prefix}{self.et*1e-6:.2f} ms"+(self.on_exit(self.et) if self.on_exit else ""))
-
-# **** tinygrad now supports dtypes! *****
-
-class DType(NamedTuple):
-  priority: int  # this determines when things get upcasted
-  itemsize: int
-  name: str
-  np: Optional[type]  # TODO: someday this will be removed with the "remove numpy" project
-  sz: int = 1
-  def __repr__(self): return f"dtypes.{self.name}"
-  @property
-  def key(self) -> str: return (self.name)
-
-# dependent typing?
-class ImageDType(DType):
-  sz = DType.sz
-
-  def __new__(cls, priority, itemsize, name, np, shape):
-    return super().__new__(cls, priority, itemsize, name, np)
-
-  def __init__(self, priority, itemsize, name, np, shape):
-    self.shape: Tuple[int, ...] = shape  # arbitrary arg for the dtype, used in image for the shape
-    super().__init__()
-
-  def __repr__(self): return f"dtypes.{self.name}({self.shape})"
-
-class dtypes:
-  @staticmethod # static methds on top, or bool in the type info will refer to dtypes.bool
-  def is_int(x: DType)-> bool: return x in (dtypes.int8, dtypes.uint8, dtypes.int32, dtypes.int64)
-  @staticmethod
-  def is_float(x: DType) -> bool: return x in (dtypes.float16, dtypes.float32, dtypes._half4, dtypes._float4)
-  @staticmethod
-  def is_unsigned(x: DType) -> bool: return x in (dtypes.uint8, dtypes.uint32, dtypes.uint64)
-  @staticmethod
-  def from_np(x) -> DType: return DTYPES_DICT[np.dtype(x).name]
-  @staticmethod
-  def fields() -> Dict[str, DType]: return DTYPES_DICT
-  bool: Final[DType] = DType(0, 1, "bool", bool)
-  float16: Final[DType] = DType(0, 2, "half", np.float16)
-  half = float16
-  float32: Final[DType] = DType(4, 4, "float", np.float32)
-  float = float32
-  int8: Final[DType] = DType(0, 1, "char", np.int8)
-  int32: Final[DType] = DType(1, 4, "int", np.int32)
-  int64: Final[DType] = DType(2, 8, "long", np.int64)
-  uint8: Final[DType] = DType(0, 1, "unsigned char", np.uint8)
-  uint32: Final[DType] = DType(1, 4, "unsigned int", np.uint32)
-  uint64: Final[DType] = DType(2, 8, "unsigned long", np.uint64)
-
-  # NOTE: these are internal dtypes, should probably check for that
-  _half4: Final[DType] = DType(0, 2*4, "half4", None, 4)
-  _float2: Final[DType] = DType(4, 4*2, "float2", None, 2)
-  _float4: Final[DType] = DType(4, 4*4, "float4", None, 4)
-
-
-DTYPES_DICT = {k: v for k, v in inspect.getmembers(dtypes()) if not k.startswith('__') and not callable(v) and not v.__class__ == staticmethod}
-
-class GlobalCounters:
-  global_ops: ClassVar[int] = 0
-  global_mem: ClassVar[int] = 0
-  time_sum_s: ClassVar[float] = 0.0
-  kernel_count: ClassVar[int] = 0
-  mem_used: ClassVar[int] = 0   # NOTE: this is not reset
-  cache: ClassVar[Optional[List[Tuple[Callable, Any]]]] = None
-  @staticmethod
-  def reset(): GlobalCounters.global_ops, GlobalCounters.global_mem, GlobalCounters.time_sum_s, GlobalCounters.kernel_count, GlobalCounters.cache = 0,0,0.0,0,None
-
-# Stripped down version of a WeakSet
-class LightWeakSet:
-  __slots__ = 'data', '_remove', '__weakref__'
-  def __init__(self):
-    self.data = set()
-    def _remove(item, selfref=ref(self)):
-      self = selfref()
-      if self: self.data.discard(item)
-    self._remove = _remove
-
-  def __len__(self): return len(self.data)
-  def add(self, item): self.data.add(ref(item, self._remove))
-  def discard(self, item): self.data.discard(ref(item))
-
-# Stripped down version of a WeakValueDictionary
-class LightWeakValueDictionary:
-  __slots__ = 'data', '_remove', '__weakref__'
-  def __init__(self):
-    def remove(wr, selfref=ref(self), _atomic_removal=_remove_dead_weakref):
-      self = selfref()
-      if self: _atomic_removal(self.data, wr.key)
-    self._remove = remove
-    self.data = {}
-
-  def __getitem__(self, key):
-    o = self.data[key]()
-    if o is None: raise KeyError(key)
-    else: return o
-
-  def __len__(self): return len(self.data)
-  def __delitem__(self, key): del self.data[key]
-  def __setitem__(self, key, value): self.data[key] = KeyedRef(value, self._remove, key)
-  def __contains__(self, key): return key in self.data
