+++ tinygrad/jit.py
@@ -1,6 +1,7 @@
 from typing import Callable, List, Tuple, Any, Dict, cast, Union
 import functools, itertools
-from tinygrad.helpers import DEBUG, DType
+from tinygrad.helpers import DEBUG
+from tinygrad.DType import DType
 
 from tinygrad.lazy import Device
 from tinygrad.tensor import Tensor
