+++ tinygrad/runtime/ops_cpu.py
@@ -1,9 +1,11 @@
 import numpy as np
 import operator
 from typing import Callable, Dict, Tuple, Optional
-from tinygrad.helpers import dtypes, DType
 from tinygrad.ops import UnaryOps, BinaryOps, MovementOps, ReduceOps, FusedOps, Op, Interpreted
 from tinygrad.runtime.lib import RawBuffer
+from tinygrad.DType import DType
+from tinygrad.dtypes import dtypes
+
 
 def shape_to_axis(old_shape:Tuple[int, ...], new_shape:Tuple[int, ...]) -> Tuple[int, ...]:
   assert len(old_shape) == len(new_shape), "reduce shapes must have same dimensions"
